+++ Editor/ProfilerWindowInterface.cs
@@ -1,15 +1,22 @@
-﻿using UnityEditor;
-using UnityEditorInternal;
+﻿using UnityEditorInternal;
 using System.Reflection;
 using System;
 using UnityEngine;
 using System.Collections.Generic;
 using UnityEngine.Profiling;
+using UnityEditor.Profiling;
 
 namespace UnityEditor.Performance.ProfileAnalyzer
 {
     internal class ProfilerWindowInterface
     {
+#if UNITY_2020_1_OR_NEWER
+        bool s_UseRawIterator = true;
+#else
+        bool s_UseRawIterator = false;
+#endif        
+        ProgressBarDisplay m_progressBar;
+
         Type m_ProfilerWindowType;
         EditorWindow m_ProfilerWindow;
         FieldInfo m_CurrentFrameFieldInfo;
@@ -27,8 +34,10 @@ namespace UnityEditor.Performance.ProfileAnalyzer
         MethodInfo m_GetProfilerModuleInfo;
         Type m_CPUProfilerModuleType;
 
-        public ProfilerWindowInterface()
+        public ProfilerWindowInterface(ProgressBarDisplay progressBar)
         {
+            m_progressBar = progressBar;
+
             Assembly assem = typeof(Editor).Assembly;
             m_ProfilerWindowType = assem.GetType("UnityEditor.ProfilerWindow");
             m_CurrentFrameFieldInfo = m_ProfilerWindowType.GetField("m_CurrentFrame", BindingFlags.NonPublic | BindingFlags.Instance);
@@ -82,19 +91,24 @@ namespace UnityEditor.Performance.ProfileAnalyzer
             return false;
         }
 
-        public bool IsProfilerWindowOpen()
+        public void GetProfilerWindowHandle()
         {
-            Profiler.BeginSample("IsProfilerWindowOpen");
+            Profiler.BeginSample("GetProfilerWindowHandle");
             UnityEngine.Object[] windows = Resources.FindObjectsOfTypeAll(m_ProfilerWindowType);
-            bool result = (windows != null && windows.Length > 0) ? true : false;
+            if (windows != null && windows.Length > 0)
+                m_ProfilerWindow = (EditorWindow)windows[0];
             Profiler.EndSample();
-
-            return result;
         }
 
         public void OpenProfilerOrUseExisting()
         {
-            m_ProfilerWindow = EditorWindow.GetWindow(m_ProfilerWindowType);
+            // Note we use existing if possible to fix a bug after domain reload
+            // Where calling EditorWindow.GetWindow directly causes a second window to open
+            if (!m_ProfilerWindow)
+            { 
+                // Create new
+                m_ProfilerWindow = EditorWindow.GetWindow(m_ProfilerWindowType);
+            }
         }
 
         public bool GetFrameRangeFromProfiler(out int first, out int last)
@@ -158,9 +172,315 @@ namespace UnityEditor.Performance.ProfileAnalyzer
 
             return null;
         }
+        public ProfileData PullFromProfiler(int firstFrameDisplayIndex, int lastFrameDisplayIndex)
+        {
+            Profiler.BeginSample("ProfilerWindowInterface.PullFromProfiler");
+            
+            bool recording = IsRecording();
+            if (recording)
+                StopRecording();
+
+            int firstFrameIndex = firstFrameDisplayIndex - 1;
+            int lastFrameIndex = lastFrameDisplayIndex - 1;
+            ProfileData profileData = GetData(firstFrameIndex, lastFrameIndex);
+
+            if (recording)
+                StartRecording();
+
+            Profiler.EndSample();
+            return profileData;
+        }
+
+#if UNITY_2020_1_OR_NEWER
+        ProfileData GetDataRaw(int firstFrameIndex, int lastFrameIndex)
+        {
+            bool firstError = true;
+
+            var data = new ProfileData();
+            data.SetFrameIndexOffset(firstFrameIndex);
+
+            var depthStack = new Stack<int>();
+
+            var threadNameCount = new Dictionary<string, int>();
+            var threadIdMapping = new Dictionary<ulong, string>();
+            var markerIdToNameIndex = new Dictionary<int, int>();
+            for (int frameIndex = firstFrameIndex; frameIndex <= lastFrameIndex; ++frameIndex)
+            {
+                m_progressBar.AdvanceProgressBar();
+
+                int threadIndex = 0;
+
+                bool threadValid = true;
+                threadNameCount.Clear();
+                ProfileFrame frame = null;
+                while (threadValid)
+                {
+                    using (RawFrameDataView frameData = ProfilerDriver.GetRawFrameDataView(frameIndex, threadIndex))
+                    {
+                        if (threadIndex == 0)
+                        {
+                            frame = new ProfileFrame();
+                            if (frameData.valid)
+                            {
+                                frame.msStartTime = frameData.frameStartTimeMs;
+                                frame.msFrame = frameData.frameTimeMs;
+                            }
+                            data.Add(frame);
+                        }
+
+                        if (!frameData.valid)
+                            break;
+
+                        string threadNameWithIndex;
+
+                        if (threadIdMapping.ContainsKey(frameData.threadId))
+                        {
+                            threadNameWithIndex = threadIdMapping[frameData.threadId];
+                        }
+                        else
+                        {
+                            string threadName = frameData.threadName;
+                            if (threadName.Trim() == "")
+                            {
+                                Debug.Log(string.Format("Warning: Unnamed thread found on frame {0}. Corrupted data suspected, ignoring frame", frameIndex));
+                                threadIndex++;
+                                continue;
+                            }
+                            var groupName = frameData.threadGroupName;
+                            threadName = ProfileData.GetThreadNameWithGroup(threadName, groupName);
+
+                            int nameCount = 0;
+                            threadNameCount.TryGetValue(threadName, out nameCount);
+                            threadNameCount[threadName] = nameCount + 1;
+
+                            threadNameWithIndex = ProfileData.ThreadNameWithIndex(threadNameCount[threadName], threadName);
+
+                            threadIdMapping[frameData.threadId] = threadNameWithIndex;
+                        }
+
+                        var thread = new ProfileThread();
+                        data.AddThreadName(threadNameWithIndex, thread);
+
+                        frame.Add(thread);
+
+                        // The markers are in depth first order 
+                        depthStack.Clear();
+                        // first sample is the thread name
+                        for (int i = 1; i < frameData.sampleCount; i++)
+                        {
+                            float durationMS = frameData.GetSampleTimeMs(i);
+                            int markerId = frameData.GetSampleMarkerId(i);
+                            if (durationMS < 0)
+                            {
+                                if (firstError)
+                                {
+                                    int displayIndex = data.OffsetToDisplayFrame(frameIndex);
+                                    string threadName = frameData.threadName;
+
+                                    string name = frameData.GetSampleName(i);
+                                    Debug.LogFormat("Ignoring Invalid marker time found for {0} on frame {1} on thread {2} ({3} < 0)",
+                                            name, displayIndex, threadName, durationMS);
+
+                                    firstError = false;
+                                }
+                            }
+                            else
+                            {
+                                int depth = 1 + depthStack.Count;
+                                var markerData = ProfileMarker.Create(durationMS, depth);
+
+                                // Use name index directly if we have already stored this named marker before
+                                int nameIndex;
+                                if (markerIdToNameIndex.TryGetValue(markerId, out nameIndex))
+                                {
+                                    markerData.nameIndex = nameIndex;
+                                }
+                                else
+                                {
+                                    string name = frameData.GetSampleName(i);
+                                    data.AddMarkerName(name, markerData);
+                                    markerIdToNameIndex[markerId] = markerData.nameIndex;
+                                }
+                                
+                                thread.Add(markerData);
+                            }
+
+                            int childrenCount = frameData.GetSampleChildrenCount(i);
+                            if (childrenCount > 0)
+                            {
+                                depthStack.Push(childrenCount);
+                            }
+                            else
+                            {
+                                while (depthStack.Count > 0)
+                                {
+                                    int remainingChildren = depthStack.Pop();
+                                    if (remainingChildren > 1)
+                                    {
+                                        depthStack.Push(remainingChildren - 1);
+                                        break;
+                                    }
+                                }
+                            }
+                        }
+                    }
+                    threadIndex++;
+                }
+            }
+
+            data.Finalise();
+
+            return data;
+        }
+#endif
+
+        ProfileData GetDataOriginal(int firstFrameIndex, int lastFrameIndex)
+        {
+            ProfilerFrameDataIterator frameData = new ProfilerFrameDataIterator();
+            bool firstError = true;
+
+            var data = new ProfileData();
+            data.SetFrameIndexOffset(firstFrameIndex);
+
+            Dictionary<string, int> threadNameCount = new Dictionary<string, int>();
+            for (int frameIndex = firstFrameIndex; frameIndex <= lastFrameIndex; ++frameIndex)
+            {
+                m_progressBar.AdvanceProgressBar();
+
+                int threadCount = frameData.GetThreadCount(frameIndex);
+                frameData.SetRoot(frameIndex, 0);
+
+                var msFrame = frameData.frameTimeMS;
+
+                /*
+                if (frameIndex == lastFrameIndex)
+                {
+                    // Check if last frame appears to be invalid data
+                    float median;
+                    float mean;
+                    float standardDeviation;
+                    CalculateFrameTimeStats(data, out median, out mean, out standardDeviation);
+                    float execessiveDeviation = (3f * standardDeviation);
+                    if (msFrame > (median + execessiveDeviation))
+                    {
+                        Debug.LogFormat("Dropping last frame as it is significantly larger than the median of the rest of the data set {0} > {1} (median {2} + 3 * standard deviation {3})", msFrame, median + execessiveDeviation, median, standardDeviation);
+                        break;
+                    }
+                    if (msFrame < (median - execessiveDeviation))
+                    {
+                        Debug.LogFormat("Dropping last frame as it is significantly smaller than the median of the rest of the data set {0} < {1} (median {2} - 3 * standard deviation {3})", msFrame, median - execessiveDeviation, median, standardDeviation);
+                        break;
+                    }
+                }
+                */
+
+                ProfileFrame frame = new ProfileFrame();
+                frame.msStartTime = 1000.0 * frameData.GetFrameStartS(frameIndex);
+                frame.msFrame = msFrame;
+                data.Add(frame);
+
+                threadNameCount.Clear();
+                for (int threadIndex = 0; threadIndex < threadCount; ++threadIndex)
+                {
+                    frameData.SetRoot(frameIndex, threadIndex);
+
+                    var threadName = frameData.GetThreadName();
+                    if (threadName.Trim() == "")
+                    {
+                        Debug.Log(string.Format("Warning: Unnamed thread found on frame {0}. Corrupted data suspected, ignoring frame", frameIndex));
+                        continue;
+                    }
+
+                    var groupName = frameData.GetGroupName();
+                    threadName = ProfileData.GetThreadNameWithGroup(threadName, groupName);
+
+                    ProfileThread thread = new ProfileThread();
+                    frame.Add(thread);
+
+                    int nameCount = 0;
+                    threadNameCount.TryGetValue(threadName, out nameCount);
+                    threadNameCount[threadName] = nameCount + 1;
+
+                    data.AddThreadName(ProfileData.ThreadNameWithIndex(threadNameCount[threadName], threadName), thread);
+
+                    const bool enterChildren = true;
+                    // The markers are in depth first order and the depth is known
+                    // So we can infer a parent child relationship
+                    while (frameData.Next(enterChildren))
+                    {
+                        if (frameData.durationMS < 0)
+                        {
+                            if (firstError)
+                            {
+                                int displayIndex = data.OffsetToDisplayFrame(frameIndex);
+
+                                Debug.LogFormat("Ignoring Invalid marker time found for {0} on frame {1} on thread {2} ({3} < 0) : Instance id : {4}",
+                                    frameData.name, displayIndex, threadName, frameData.durationMS, frameData.instanceId);
+
+                                firstError = false;
+                            }
+                            continue;
+                        }
+                        var markerData = ProfileMarker.Create(frameData);
+
+                        data.AddMarkerName(frameData.name, markerData);
+                        thread.Add(markerData);
+                    }
+                }
+            }
+
+            data.Finalise();
+
+            frameData.Dispose();
+            return data;
+        }
+
+
+        ProfileData GetData(int firstFrameIndex, int lastFrameIndex)
+        {
+            //s_UseRawIterator ^= true;
+
+            System.Diagnostics.Stopwatch stopwatch = new System.Diagnostics.Stopwatch();
+            stopwatch.Start();
+
+            ProfileData data;
+#if UNITY_2020_1_OR_NEWER
+            if (s_UseRawIterator)
+            {
+                data = GetDataRaw(firstFrameIndex, lastFrameIndex);
+            }
+            else
+#endif
+            {
+                data = GetDataOriginal(firstFrameIndex, lastFrameIndex);
+            }
+
+            stopwatch.Stop();
+            //Debug.LogFormat("Pull time {0}ms ({1})", stopwatch.ElapsedMilliseconds, s_UseRawIterator ? "Raw" : "Standard");
+           
+            return data;
+        }
+
+#if UNITY_2020_1_OR_NEWER
+        public float GetFrameTimeRaw(int frameIndex)
+        {
+            using (RawFrameDataView frameData = ProfilerDriver.GetRawFrameDataView(frameIndex, 0))
+            {
+                if (!frameData.valid)
+                    return 0f;
+
+                return frameData.frameTimeMs;
+            }
+        }
+#endif
+
 
         public float GetFrameTime(int frameIndex)
         {
+#if UNITY_2020_1_OR_NEWER
+            if (s_UseRawIterator)
+                return GetFrameTimeRaw(frameIndex);
+#endif
             ProfilerFrameDataIterator frameData = new ProfilerFrameDataIterator();
 
             frameData.SetRoot(frameIndex, 0);
@@ -309,6 +629,22 @@ namespace UnityEditor.Performance.ProfileAnalyzer
             return ProfilerDriver.enabled;
 #else
             return false;
+#endif
+        }
+
+        public void StopRecording()
+        {
+#if UNITY_2017_4_OR_NEWER
+            // Stop recording first
+            ProfilerDriver.enabled = false;
+#endif
+        }
+
+        public void StartRecording()
+        {
+#if UNITY_2017_4_OR_NEWER
+            // Stop recording first
+            ProfilerDriver.enabled = true;
 #endif
         }
     }